pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.isort]
profile = "black"
skip = ["env", "venv"]

[tool.poetry]
authors = ["Alex"]
description = "Playwright Python Airbnb with pytest and Allure report"
keywords = ["playwright", "automation", "testing", "web"]
name = "playwright-python"
packages = [{include = "playwright_python"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
allure-pytest = "2.13.5"
axe-playwright-python = "0.1.4"
playwright = "1.50.0"
pytest = "8.3.4"
pytest-base-url = "2.1.0"
pytest-playwright = "0.7.0"
pytest-split = "0.10.0"
python = "^3.11"
requests = "2.32.3"

[tool.poetry.group.dev.dependencies]
black = "25.1.0"
isort = "6.0.1"
pre-commit = "4.1.0"

[tool.pytest.ini_options]
addopts = [
  "--clean-alluredir",
  "--alluredir=allure-results",
  "--allure-link-pattern",
  "issue:https://{}",
  "--allure-link-pattern",
  "link:https://{}",
  "--allure-link-pattern",
  "tms:https://{}",
  "-ra",
  "--color=yes",
  "--browser-channel=chrome",
  "--headed",
  "--video=retain-on-failure",
  "--tracing=retain-on-failure"
]
base_url = 'https://www.airbnb.com/'
log_cli = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_level = "INFO"
markers = [
  "airbnb_sanity: quick sanity on main path of reservation"
]
